// src/api/api-course/models/DayDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project & Day API
 * An OpenAPI specification for project management, day templates, tasks, forms, media, and related user progress tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DayBlueprint } from './DayBlueprint';
import {
    DayBlueprintFromJSON,
    DayBlueprintFromJSONTyped,
    DayBlueprintToJSON,
    DayBlueprintToJSONTyped,
} from './DayBlueprint';
import type { DayInstance } from './DayInstance';
import {
    DayInstanceFromJSON,
    DayInstanceFromJSONTyped,
    DayInstanceToJSON,
    DayInstanceToJSONTyped,
} from './DayInstance';

/**
 * 
 * @export
 * @interface DayDTO
 */
export interface DayDTO {
    /**
     * 
     * @type {DayBlueprint}
     * @memberof DayDTO
     */
    blueprint: DayBlueprint;
    /**
     * 
     * @type {DayInstance}
     * @memberof DayDTO
     */
    instance: DayInstance;
}

/**
 * Check if a given object implements the DayDTO interface.
 */
export function instanceOfDayDTO(value: object): value is DayDTO {
    if (!('blueprint' in value) || value['blueprint'] === undefined) return false;
    if (!('instance' in value) || value['instance'] === undefined) return false;
    return true;
}

export function DayDTOFromJSON(json: any): DayDTO {
    return DayDTOFromJSONTyped(json, false);
}

export function DayDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DayDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'blueprint': DayBlueprintFromJSON(json['blueprint']),
        'instance': DayInstanceFromJSON(json['instance']),
    };
}

export function DayDTOToJSON(json: any): DayDTO {
    return DayDTOToJSONTyped(json, false);
}

export function DayDTOToJSONTyped(value?: DayDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'blueprint': DayBlueprintToJSON(value['blueprint']),
        'instance': DayInstanceToJSON(value['instance']),
    };
}