// src/api/api-course/models/DayComponentsMapper.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project & Day API
 * An OpenAPI specification for project management, day templates, tasks, forms, media, and related user progress tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DayComponentsMapperId } from './DayComponentsMapperId';
import {
    DayComponentsMapperIdFromJSON,
    DayComponentsMapperIdFromJSONTyped,
    DayComponentsMapperIdToJSON,
    DayComponentsMapperIdToJSONTyped,
} from './DayComponentsMapperId';

/**
 * Mapping of components (tasks, forms, media) within a day
 * @export
 * @interface DayComponentsMapper
 */
export interface DayComponentsMapper {
    /**
     * 
     * @type {DayComponentsMapperId}
     * @memberof DayComponentsMapper
     */
    id: DayComponentsMapperId;
    /**
     * 
     * @type {string}
     * @memberof DayComponentsMapper
     */
    type: DayComponentsMapperTypeEnum;
    /**
     * Order or position within a list
     * @type {number}
     * @memberof DayComponentsMapper
     */
    sortOrder: number;
}


/**
 * @export
 */
export const DayComponentsMapperTypeEnum = {
    Task: 'TASK',
    Form: 'FORM',
    Media: 'MEDIA'
} as const;
export type DayComponentsMapperTypeEnum = typeof DayComponentsMapperTypeEnum[keyof typeof DayComponentsMapperTypeEnum];


/**
 * Check if a given object implements the DayComponentsMapper interface.
 */
export function instanceOfDayComponentsMapper(value: object): value is DayComponentsMapper {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('sortOrder' in value) || value['sortOrder'] === undefined) return false;
    return true;
}

export function DayComponentsMapperFromJSON(json: any): DayComponentsMapper {
    return DayComponentsMapperFromJSONTyped(json, false);
}

export function DayComponentsMapperFromJSONTyped(json: any, ignoreDiscriminator: boolean): DayComponentsMapper {
    if (json == null) {
        return json;
    }
    return {
        
        'id': DayComponentsMapperIdFromJSON(json['id']),
        'type': json['type'],
        'sortOrder': json['sortOrder'],
    };
}

export function DayComponentsMapperToJSON(json: any): DayComponentsMapper {
    return DayComponentsMapperToJSONTyped(json, false);
}

export function DayComponentsMapperToJSONTyped(value?: DayComponentsMapper | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': DayComponentsMapperIdToJSON(value['id']),
        'type': value['type'],
        'sortOrder': value['sortOrder'],
    };
}