// src/api/api-content/models/TaskDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project & Day API
 * An OpenAPI specification for project management, day templates, tasks, forms, media, and related user progress tracking.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TaskInstance } from './TaskInstance';
import {
    TaskInstanceFromJSON,
    TaskInstanceFromJSONTyped,
    TaskInstanceToJSON,
    TaskInstanceToJSONTyped,
} from './TaskInstance';
import type { TaskBlueprint } from './TaskBlueprint';
import {
    TaskBlueprintFromJSON,
    TaskBlueprintFromJSONTyped,
    TaskBlueprintToJSON,
    TaskBlueprintToJSONTyped,
} from './TaskBlueprint';

/**
 * 
 * @export
 * @interface TaskDTO
 */
export interface TaskDTO {
    /**
     * 
     * @type {TaskBlueprint}
     * @memberof TaskDTO
     */
    blueprint: TaskBlueprint;
    /**
     * 
     * @type {TaskInstance}
     * @memberof TaskDTO
     */
    progress: TaskInstance;
}

/**
 * Check if a given object implements the TaskDTO interface.
 */
export function instanceOfTaskDTO(value: object): value is TaskDTO {
    if (!('blueprint' in value) || value['blueprint'] === undefined) return false;
    if (!('progress' in value) || value['progress'] === undefined) return false;
    return true;
}

export function TaskDTOFromJSON(json: any): TaskDTO {
    return TaskDTOFromJSONTyped(json, false);
}

export function TaskDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): TaskDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'blueprint': TaskBlueprintFromJSON(json['blueprint']),
        'progress': TaskInstanceFromJSON(json['progress']),
    };
}

export function TaskDTOToJSON(json: any): TaskDTO {
    return TaskDTOToJSONTyped(json, false);
}

export function TaskDTOToJSONTyped(value?: TaskDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'blueprint': TaskBlueprintToJSON(value['blueprint']),
        'progress': TaskInstanceToJSON(value['progress']),
    };
}